Implement Slack client configuration and handlers; add environment variable validation and server setup
