+++ src/schemas/index.ts
@@ -1,0 +1,2 @@
+// スキーマをすべてエクスポート
+export * from '../schemas.js';
